wip: Try to fix Heizkostenabrechnung

+++ Deeplex.Saverwalter.WebAPI/Services/DbServices/UmlageDbService.cs
@@ -180,7 +180,11 @@ namespace Deeplex.Saverwalter.WebAPI.Services.ControllerService
             {
                 if (hkvo.Id == 0)
                 {
-                    var newHKVO = new HKVO(hkvo.HKVO_P7 / 100, hkvo.HKVO_P8 / 100, (HKVO_P9A2)hkvo.HKVO_P9.Id, hkvo.Strompauschale / 100)
+                    var newHKVO = new HKVO(
+                        ((double)hkvo.HKVO_P7) / 100,
+                        ((double)hkvo.HKVO_P8) / 100,
+                        (HKVO_P9A2)hkvo.HKVO_P9.Id,
+                        ((double)hkvo.Strompauschale) / 100)
                     {
                         Betriebsstrom = Ctx.Umlagen.Single(e => e.UmlageId == hkvo.Stromrechnung.Id)
                     };
@@ -190,10 +194,10 @@ namespace Deeplex.Saverwalter.WebAPI.Services.ControllerService
                 else
                 {
                     var oldHKVO = Ctx.HKVO.Single(e => e.HKVOId == hkvo.Id);
-                    oldHKVO.HKVO_P7 = (double)hkvo.HKVO_P7 / 100;
-                    oldHKVO.HKVO_P8 = (double)hkvo.HKVO_P8 / 100;
+                    oldHKVO.HKVO_P7 = ((double)hkvo.HKVO_P7) / 100;
+                    oldHKVO.HKVO_P8 = ((double)hkvo.HKVO_P8) / 100;
                     oldHKVO.HKVO_P9 = (HKVO_P9A2)hkvo.HKVO_P9.Id;
-                    oldHKVO.Strompauschale = (double)hkvo.Strompauschale / 100;
+                    oldHKVO.Strompauschale = ((double)hkvo.Strompauschale) / 100;
                     oldHKVO.Betriebsstrom = Ctx.Umlagen.Single(e => e.UmlageId == hkvo.Stromrechnung.Id);
                     Ctx.HKVO.Update(oldHKVO);
                 }
